# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Farm Idle Game API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface FarmState {
  wheat: number;
  chicken: number;
  cow: number;
}

export interface User {
  telegramId: string;
  firstName: string;
  /** @nullable */
  username?: string | null;
  coins: number;
  balance: number;
  farmState?: FarmState;
  streakCount: number;
  totalReferrals: number;
  /** @nullable */
  lastSpinAt: string | null;
  createdAt: string;
}

export interface UserInitInput {
  telegramId: string;
  firstName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  referredBy?: string | null;
}

export interface FarmStateInput {
  balance: number;
  farmState: FarmState;
}

export interface SpinInput {
  telegramId: string;
}

export type SpinResultPrizeType = typeof SpinResultPrizeType[keyof typeof SpinResultPrizeType];


export const SpinResultPrizeType = {
  coins: 'coins',
  boost: 'boost',
  jackpot: 'jackpot',
  miss: 'miss',
} as const;

export interface SpinResult {
  prize: string;
  prizeType: SpinResultPrizeType;
  coinsEarned: number;
  newCoinsTotal: number;
  segmentIndex: number;
  canSpinAgainAt: string;
  /** @nullable */
  streakBonus?: number | null;
}

export interface SpinCooldown {
  canSpinAgainAt: string;
  hoursRemaining: number;
}

export interface LeaderboardEntry {
  rank: number;
  telegramId: string;
  firstName: string;
  /** @nullable */
  username?: string | null;
  coins: number;
  totalReferrals: number;
}

export interface Achievement {
  key: string;
  title: string;
  description: string;
  icon: string;
  earnedAt: string;
}

export interface StarPackage {
  id: string;
  stars: number;
  coinsRequired: number;
  label: string;
  popular: boolean;
}

export interface StarPurchaseInput {
  telegramId: string;
  packageId: string;
}

export interface StarPurchaseResult {
  success: boolean;
  invoiceLink: string;
  starsAwarded: number;
  coinsSpent: number;
  newCoinsTotal: number;
}

export interface ReferralEntry {
  firstName: string;
  joinedAt: string;
  coinsEarned: number;
}

export interface ReferralStats {
  telegramId: string;
  totalReferrals: number;
  coinsFromReferrals: number;
  referralLink: string;
  recentReferrals: ReferralEntry[];
}

export interface WithdrawInput {
  telegramId: string;
  /** Amount in TL to withdraw (min 50, max 350) */
  amount: number;
  /** Withdrawal method: papara, iban, crypto */
  method: string;
}

export type WithdrawResultStatus = typeof WithdrawResultStatus[keyof typeof WithdrawResultStatus];


export const WithdrawResultStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface WithdrawResult {
  success: boolean;
  requestId: string;
  amount: number;
  status: WithdrawResultStatus;
  message: string;
  newBalance: number;
}

export interface WithdrawEntry {
  id: string;
  amount: number;
  method: string;
  status: string;
  createdAt: string;
}

export interface OnlineStats {
  onlineCount: number;
  totalPlayers: number;
  totalCoinsInCirculation: number;
}

export interface VaultDepositInput {
  telegramId: string;
  /** Number of coins to deposit (min 100) */
  coins: number;
  /** Lock period in days: 3, 7, 14, or 30 */
  lockDays: number;
}

export interface VaultDepositResult {
  id: string;
  coinsDeposited: number;
  coinsToReceive: number;
  multiplier: number;
  maturesAt: string;
  lockDays: number;
}

export type VaultEntryStatus = typeof VaultEntryStatus[keyof typeof VaultEntryStatus];


export const VaultEntryStatus = {
  locked: 'locked',
  mature: 'mature',
  claimed: 'claimed',
} as const;

export interface VaultEntry {
  id: string;
  coinsDeposited: number;
  coinsToReceive: number;
  multiplier: number;
  lockDays: number;
  maturesAt: string;
  status: VaultEntryStatus;
  createdAt?: string;
}

export interface VaultClaimInput {
  telegramId: string;
  depositId: string;
}

export interface VaultClaimResult {
  coinsReceived: number;
  newCoinTotal: number;
}

export type NftItemRarity = typeof NftItemRarity[keyof typeof NftItemRarity];


export const NftItemRarity = {
  common: 'common',
  rare: 'rare',
  special: 'special',
  legendary: 'legendary',
} as const;

export interface NftItem {
  id: string;
  ownerTelegramId: string;
  nftType: string;
  rarity: NftItemRarity;
  name: string;
  emoji: string;
  mintNumber: number;
  isListedForTrade: boolean;
  createdAt: string;
}

export interface NftListTradeInput {
  telegramId: string;
  nftId: string;
  list: boolean;
}

export interface NftTradeOfferInput {
  offererTelegramId: string;
  offeredNftId: string;
  targetTelegramId?: string;
  wantedNftType?: string;
}

export type NftTradeOfferStatus = typeof NftTradeOfferStatus[keyof typeof NftTradeOfferStatus];


export const NftTradeOfferStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export interface NftTradeOffer {
  id: string;
  offererTelegramId: string;
  offeredNftId: string;
  targetTelegramId?: string;
  wantedNftType?: string;
  status: NftTradeOfferStatus;
  createdAt: string;
  offeredNft?: NftItem;
}

export interface NftTradeAcceptInput {
  telegramId: string;
  offerId: string;
  acceptorNftId: string;
}

export type GetLeaderboardParams = {
limit?: number;
};

export type AcceptTradeOffer200 = {
  success?: boolean;
};