# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Farm Idle Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nfts/user/{telegramId}:
    get:
      operationId: getUserNfts
      tags: [nfts]
      summary: Get all NFTs owned by a user
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User NFTs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NftItem"
  /nfts/list-trade:
    post:
      operationId: listNftForTrade
      tags: [nfts]
      summary: List or delist an NFT for trading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NftListTradeInput"
      responses:
        "200":
          description: NFT updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NftItem"
...
  /nfts/trade/offer:
    post:
      operationId: createTradeOffer
      tags: [nfts]
      summary: Create a trade offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NftTradeOfferInput"
      responses:
        "200":
          description: Trade offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NftTradeOffer"
  /nfts/trade/offers/{telegramId}:
    get:
      operationId: getTradeOffers
      tags: [nfts]
      summary: Get incoming trade offers for a user
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NftTradeOffer"
  /nfts/trade/accept:
    post:
      operationId: acceptTradeOffer
      tags: [nfts]
      summary: Accept a trade offer (swap NFTs)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NftTradeAcceptInput"
      responses:
        "200":
          description: Trade completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
...
    NftItem:
      type: object
      required: [id, ownerTelegramId, nftType, rarity, name, emoji, mintNumber, isListedForTrade, createdAt]
      properties:
        id:
          type: string
        ownerTelegramId:
          type: string
        nftType:
          type: string
        rarity:
          type: string
          enum: [common, rare, special, legendary]
        name:
          type: string
        emoji:
          type: string
        mintNumber:
          type: integer
        isListedForTrade:
          type: boolean
        createdAt:
          type: string
    NftListTradeInput:
      type: object
      required: [telegramId, nftId, list]
      properties:
        telegramId:
          type: string
        nftId:
          type: string
        list:
          type: boolean
    NftTradeOfferInput:
      type: object
      required: [offererTelegramId, offeredNftId]
      properties:
        offererTelegramId:
          type: string
        offeredNftId:
          type: string
        targetTelegramId:
          type: string
        wantedNftType:
          type: string
    NftTradeOffer:
      type: object
      required: [id, offererTelegramId, offeredNftId, status, createdAt]
      properties:
        id:
          type: string
        offererTelegramId:
          type: string
        offeredNftId:
          type: string
        targetTelegramId:
          type: string
        wantedNftType:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled]
        createdAt:
          type: string
        offeredNft:
          $ref: "#/components/schemas/NftItem"
    NftTradeAcceptInput:
      type: object
      required: [telegramId, offerId, acceptorNftId]
      properties:
        telegramId:
          type: string
        offerId:
          type: string
        acceptorNftId:
          type: string