# lib/api-zod/src/generated/types/nftItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Farm Idle Game API
 * OpenAPI spec version: 0.3.0
 */
import type { NftItemRarity } from './nftItemRarity';

export interface NftItem {
  id: string;
  ownerTelegramId: string;
  nftType: string;
  rarity: NftItemRarity;
  name: string;
  emoji: string;
  mintNumber: number;
  isListedForTrade: boolean;
  createdAt: string;
}