# lib/api-zod/src/generated/types/nftTradeOffer.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Farm Idle Game API
 * OpenAPI spec version: 0.3.0
 */
import type { NftItem } from './nftItem';
import type { NftTradeOfferStatus } from './nftTradeOfferStatus';

export interface NftTradeOffer {
  id: string;
  offererTelegramId: string;
  offeredNftId: string;
  targetTelegramId?: string;
  wantedNftType?: string;
  status: NftTradeOfferStatus;
  createdAt: string;
  offeredNft?: NftItem;
}