# lib/api-zod/src/generated/types/nftTradeOfferInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Farm Idle Game API
 * OpenAPI spec version: 0.3.0
 */

export interface NftTradeOfferInput {
  offererTelegramId: string;
  offeredNftId: string;
  targetTelegramId?: string;
  wantedNftType?: string;
}